added named slots for flexible use of child elements
this introduces a convention to extend macros' signature with named
blocks as child elements, which can then be re-arranged by the macro -
while technically we could also use parameters for this, passing child
elements that way is awkward

addressing this use case (cf. `Card` example in tests) would otherwise
require a DOM-style API for introspection and manipulation, which is
problematic in various ways: it requires some sort of AST, risks
breaking macros' encapsulation etc.

we're taking advantage of the fact that macros can return arbitrary
children, so `<Slot>` generates a (crude) wrapper object which can then
be identified by the respective macro

NB: this is an optional runtime feature provided mostly for convenience;
    `slots.js` might just as well be implemented in userland